refactor: delete sharedGameController move sharedGameController method
